backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.3
gunicorn==22.0.0
requests==2.31.0
sendgrid==6.11.0
